+++ src/test/groovy/ru/vyarus/guice/persist/orient/repository/command/function/FunctionRemoteTest.groovy
@@ -1,5 +1,9 @@
 package ru.vyarus.guice.persist.orient.repository.command.function
 
+import com.orientechnologies.orient.core.sql.OCommandSQL
+import ru.vyarus.guice.persist.orient.db.transaction.template.SpecificTxAction
+import ru.vyarus.guice.persist.orient.repository.core.MethodExecutionException
+import ru.vyarus.guice.persist.orient.support.model.Model
 import ru.vyarus.guice.persist.orient.util.remoteext.UseRemote
 
 /**
@@ -8,4 +12,19 @@ import ru.vyarus.guice.persist.orient.util.remoteext.UseRemote
  */
 @UseRemote
 class FunctionRemoteTest extends FunctionRecognitionTest {
+
+    // remove this test and enable root class test if bug would be fixed
+    def "Check function calls"() {
+        context.doInTransaction({ db ->
+            db.save(new Model(name: 'John', nick: 'Doe'))
+            db.command(new OCommandSQL("CREATE FUNCTION function1 \"select from Model\" LANGUAGE SQL ")).execute();
+        } as SpecificTxAction)
+
+        when: "calling function"
+        List<Model> res = dao.function()
+        // todo this is orient bug in legacy api
+        then:
+        def ex = thrown(MethodExecutionException)
+        ex.printStackTrace()
+    }
 }